Куча изменений давно не обновлял

+++ GKHNNC/Controllers/OsmotrWorksController.cs
@@ -18,8 +18,10 @@ namespace GKHNNC.Controllers
         // GET: OsmotrWorks
         public ActionResult Index()
         {
-            var osmotrWorks = db.OsmotrWorks.Include(o => o.DOMPart).Include(o => o.Izmerenie);
-            return View(osmotrWorks.ToList());
+            var osmotrWorks = db.OsmotrWorks.Include(o => o.DOMPart).Include(o => o.Izmerenie).OrderBy(x=>x.DOMPartId).ThenBy(x=>x.Name).ToList();
+            List<SelectListItem> OW = db.OsmotrWorks.OrderBy(x => x.DOMPartId).ThenBy(x=>x.Name).Select(a => new SelectListItem { Value = a.Id.ToString(), Text =a.DOMPart.Name+" "+ a.Name }).ToList();
+            ViewBag.OW = OW;
+            return View(osmotrWorks);
         }
 
         // GET: OsmotrWorks/Details/5
@@ -50,14 +52,14 @@ namespace GKHNNC.Controllers
         // сведения см. в статье https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Name,IzmerenieId,Cost,DOMPartId")] OsmotrWork osmotrWork)
+        public ActionResult Create([Bind(Include = "Id,Name,IzmerenieId,Cost,DOMPartId,OtchetId")] OsmotrWork osmotrWork)
         {
-            if (ModelState.IsValid)
-            {
+          //  if (ModelState.IsValid)
+         //   {
                 db.OsmotrWorks.Add(osmotrWork);
                 db.SaveChanges();
                 return RedirectToAction("Index");
-            }
+          //  }
 
             ViewBag.DOMPartId = new SelectList(db.DOMParts, "Id", "Name", osmotrWork.DOMPartId);
             ViewBag.IzmerenieId = new SelectList(db.Izmerenies, "Id", "Name", osmotrWork.IzmerenieId);
@@ -86,7 +88,7 @@ namespace GKHNNC.Controllers
         // сведения см. в статье https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Name,IzmerenieId,Cost,DOMPartId")] OsmotrWork osmotrWork)
+        public ActionResult Edit([Bind(Include = "Id,Name,IzmerenieId,Cost,DOMPartId,OtchetId")] OsmotrWork osmotrWork)
         {
             if (ModelState.IsValid)
             {
@@ -113,6 +115,45 @@ namespace GKHNNC.Controllers
             }
             return View(osmotrWork);
         }
+        [HttpPost]
+        public JsonResult Count(int? id)
+        {
+            if (id == null)
+            {
+                return Json("Error");
+            }
+            OsmotrWork osmotrWork = db.OsmotrWorks.Find(id);
+            int Col = db.ActiveOsmotrWorks.Where(x => x.OsmotrWorkId == id).Count();
+            
+            return Json("Таких активных работ в осмотрах "+Col);
+        }
+        [HttpPost]
+        public JsonResult Replace(int? FromId,int?ToId)
+        {
+            if (FromId == null&&ToId==null)
+            {
+                return Json("Error");
+            }
+      
+            List<ActiveOsmotrWork> From = db.ActiveOsmotrWorks.Where(x => x.OsmotrWorkId == FromId).ToList();
+
+            int Count = 0;
+            foreach (ActiveOsmotrWork F in From)
+            {
+                try
+                {
+                    F.OsmotrWorkId = ToId.Value;
+                    db.Entry(F).State = EntityState.Modified;
+                    db.SaveChanges();
+                    Count++;
+                }
+                catch
+                {
+                    return Json("Error");
+                }
+            }
+            return Json("Произведено замен " + Count);
+        }
 
         // POST: OsmotrWorks/Delete/5
         [HttpPost, ActionName("Delete")]
@@ -120,8 +161,17 @@ namespace GKHNNC.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             OsmotrWork osmotrWork = db.OsmotrWorks.Find(id);
-            db.OsmotrWorks.Remove(osmotrWork);
-            db.SaveChanges();
+           
+            int Col = db.ActiveOsmotrWorks.Where(x => x.OsmotrWorkId == id).Count();
+            if (Col == 0)
+            {
+                db.OsmotrWorks.Remove(osmotrWork);
+                db.SaveChanges();
+            }
+            else
+            {
+                return Json("Работы не пустые!!!");
+            }
             return RedirectToAction("Index");
         }
